refactor: separate unit tests from integration tests in powerups
- Move integration tests to dedicated integration_test.go files
- Keep only pure unit tests (Action generation) in main test files
- Integration tests include: environment manipulation, filesystem I/O, registry usage
- Unit tests verify PowerUp -> Action transformation as pure functions
- Aligns with original design: PowerUps as pure functions, synthfs for mutations

This separation makes tests:
- Faster to run (unit tests don't touch filesystem)
- Easier to understand (clear separation of concerns)
- More maintainable (obvious what each test is verifying)
