+++ pkg/powerups/symlink/symlink_test.go
@@ -6,7 +6,6 @@ import (
 	"testing"
 
 	"github.com/arthur-debert/dodot/pkg/config"
-	"github.com/arthur-debert/dodot/pkg/registry"
 	"github.com/arthur-debert/dodot/pkg/types"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -129,39 +128,6 @@ func TestSymlinkPowerUp_ConflictDetection(t *testing.T) {
 	assert.Contains(t, err.Error(), "/dotfiles/pack2/.config")
 }
 
-func TestSymlinkPowerUp_EnvironmentExpansion(t *testing.T) {
-	oldCustomDir := os.Getenv("CUSTOM_DIR")
-	require.NoError(t, os.Setenv("CUSTOM_DIR", "/expanded/path"))
-	defer func() {
-		if oldCustomDir != "" {
-			require.NoError(t, os.Setenv("CUSTOM_DIR", oldCustomDir))
-		} else {
-			require.NoError(t, os.Unsetenv("CUSTOM_DIR"))
-		}
-	}()
-
-	powerUp := NewSymlinkPowerUp()
-
-	matches := []types.TriggerMatch{
-		{
-			TriggerName:  "filename",
-			Pack:         "test",
-			Path:         "file.txt",
-			AbsolutePath: "/dotfiles/test/file.txt",
-			PowerUpName:  "symlink",
-			PowerUpOptions: map[string]interface{}{
-				"target": "$CUSTOM_DIR/configs",
-			},
-		},
-	}
-
-	actions, err := powerUp.Process(matches)
-	require.NoError(t, err)
-	require.Len(t, actions, 1)
-
-	assert.Equal(t, "/expanded/path/configs/file.txt", actions[0].Target)
-}
-
 func TestSymlinkPowerUp_ValidateOptions(t *testing.T) {
 	powerUp := NewSymlinkPowerUp()
 
@@ -250,20 +216,6 @@ func TestSymlinkPowerUp_MetadataInActions(t *testing.T) {
 
 	// Check that trigger name is preserved in action metadata
 	assert.Equal(t, "glob", actions[0].Metadata["trigger"])
-}
-
-func TestSymlinkPowerUp_FactoryRegistration(t *testing.T) {
-	// Test that the factory is registered
-	factory, err := registry.GetPowerUpFactory(SymlinkPowerUpName)
-	require.NoError(t, err)
-	require.NotNil(t, factory)
-
-	// Test factory creates power-up correctly
-	powerUp, err := factory(nil)
-	require.NoError(t, err)
-	require.NotNil(t, powerUp)
-
-	assert.Equal(t, SymlinkPowerUpName, powerUp.Name())
 }
 
 func TestSymlinkPowerUp_PreservesDirectoryStructure(t *testing.T) {
